+++ plugins/source/aws/client/services.go
@@ -16,6 +16,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/athena"
 	"github.com/aws/aws-sdk-go-v2/service/autoscaling"
 	"github.com/aws/aws-sdk-go-v2/service/backup"
+	"github.com/aws/aws-sdk-go-v2/service/batch"
 	"github.com/aws/aws-sdk-go-v2/service/cloudformation"
 	"github.com/aws/aws-sdk-go-v2/service/cloudfront"
 	"github.com/aws/aws-sdk-go-v2/service/cloudhsmv2"
@@ -26,6 +27,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/codepipeline"
 	"github.com/aws/aws-sdk-go-v2/service/cognitoidentity"
 	"github.com/aws/aws-sdk-go-v2/service/cognitoidentityprovider"
+	"github.com/aws/aws-sdk-go-v2/service/computeoptimizer"
 	"github.com/aws/aws-sdk-go-v2/service/configservice"
 	"github.com/aws/aws-sdk-go-v2/service/databasemigrationservice"
 	"github.com/aws/aws-sdk-go-v2/service/dax"
@@ -122,6 +124,7 @@ func initServices(region string, c aws.Config) Services {
 		Appsync:                   appsync.NewFromConfig(awsCfg),
 		Athena:                    athena.NewFromConfig(awsCfg),
 		Autoscaling:               autoscaling.NewFromConfig(awsCfg),
+		Batch:                     batch.NewFromConfig(awsCfg),
 		Backup:                    backup.NewFromConfig(awsCfg),
 		Cloudformation:            cloudformation.NewFromConfig(awsCfg),
 		Cloudfront:                cloudfront.NewFromConfig(awsCfg),
@@ -133,6 +136,7 @@ func initServices(region string, c aws.Config) Services {
 		Codepipeline:              codepipeline.NewFromConfig(awsCfg),
 		Cognitoidentity:           cognitoidentity.NewFromConfig(awsCfg),
 		Cognitoidentityprovider:   cognitoidentityprovider.NewFromConfig(awsCfg),
+		Computeoptimizer:          computeoptimizer.NewFromConfig(awsCfg),
 		Configservice:             configservice.NewFromConfig(awsCfg),
 		Databasemigrationservice:  databasemigrationservice.NewFromConfig(awsCfg),
 		Dax:                       dax.NewFromConfig(awsCfg),
@@ -227,6 +231,7 @@ type Services struct {
 	Athena                    services.AthenaClient
 	Autoscaling               services.AutoscalingClient
 	Backup                    services.BackupClient
+	Batch                     services.BatchClient
 	Cloudformation            services.CloudformationClient
 	Cloudfront                services.CloudfrontClient
 	Cloudhsmv2                services.Cloudhsmv2Client
@@ -237,6 +242,7 @@ type Services struct {
 	Codepipeline              services.CodepipelineClient
 	Cognitoidentity           services.CognitoidentityClient
 	Cognitoidentityprovider   services.CognitoidentityproviderClient
+	Computeoptimizer          services.ComputeoptimizerClient
 	Configservice             services.ConfigserviceClient
 	Databasemigrationservice  services.DatabasemigrationserviceClient
 	Dax                       services.DaxClient